API refactoring
