--- api/requirements.txt
@@ -1,44 +1,0 @@
-anyio==3.6.2
-asyncio==3.4.3
-beanie==1.17.0
-certifi==2022.12.7
-charset-normalizer==3.1.0
-click==8.1.3
-dnspython==2.3.0
-email-validator==1.3.1
-fastapi==0.95.0
-filelock==3.10.2
-h11==0.14.0
-httpcore==0.16.3
-httptools==0.5.0
-httpx==0.23.3
-huggingface-hub==0.13.3
-idna==3.4
-itsdangerous==2.1.2
-Jinja2==3.1.2
-lazy-model==0.0.5
-MarkupSafe==2.1.2
-motor==3.1.1
-orjson==3.8.8
-packaging==23.0
-psutil==5.9.4
-pydantic==1.10.7
-pymongo==4.3.3
-python-dotenv==1.0.0
-python-multipart==0.0.6
-PyYAML==6.0
-requests==2.28.2
-rfc3986==1.5.0
-sentencepiece==0.1.97
-sniffio==1.3.0
-sse-starlette==1.3.3
-starlette==0.26.1
-toml==0.10.2
-tqdm==4.65.0
-typing_extensions==4.5.0
-ujson==5.7.0
-urllib3==1.26.15
-uvicorn==0.21.1
-uvloop==0.17.0
-watchfiles==0.18.1
-websockets==10.4
