+++ scripts/compile.sh
